% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{coxPH_plot}
\alias{coxPH_plot}
\title{Compute KDE and JSD

Compute kernel density estimate (KDE) of a marker expression and
Jensen Shannon distance between two computed KDEs}
\usage{
coxPH_plot(x)
}
\arguments{
\item{x}{is a list of marker expression values in different cells of a patient}

\item{m}{is the number of grids used in KDE, default is m = 1024}

\item{px}{is the KDE of first marker}

\item{py}{is the KDE of second marker}
}
\value{
The function dens_univ returns KDE and the function jensen_shannon_dist returns the JSD between two densities
}
\description{
Compute KDE and JSD

Compute kernel density estimate (KDE) of a marker expression and
Jensen Shannon distance between two computed KDEs
}
