% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_computation.R
\name{dens_univ}
\alias{dens_univ}
\title{Compute KDE and JSD
Compute kernel density estimate (KDE) of a marker expression using the function, dens_univ,
compute Jensen Shannon Distance (JSD) between two computed KDEs using the function, jensen_shannon_dist,
compute the KDE of all the patients of the dataset using the function, Array_KDE, and
compute Jensen Shannon Distance matrix between all the patients using the function, JSD_matrix.}
\usage{
dens_univ(x, ngrids = 1024, min_coef = 0, max_coef = 1)
}
\arguments{
\item{x}{is a list of marker expression values in different cells of a patient}

\item{ngrids}{is the number of grids used in KDE, default is m = 1024}

\item{px}{is the KDE of first marker}

\item{py}{is the KDE of second marker}

\item{Data}{is the dataset having one column named "SampleID" with the patient IDs and one column with marker expression values}
}
\value{
The function, dens_univ returns KDE and the grid-points as a list, the function, jensen_shannon_dist returns the JSD
between two densities the function, Array_KDE returns KDE (and the grid-points) of all the patients in a form of a 3d array,
and the function, JSD_matrix returns the JSD distance between all the images in a matrix form.
}
\description{
Compute KDE and JSD
Compute kernel density estimate (KDE) of a marker expression using the function, dens_univ,
compute Jensen Shannon Distance (JSD) between two computed KDEs using the function, jensen_shannon_dist,
compute the KDE of all the patients of the dataset using the function, Array_KDE, and
compute Jensen Shannon Distance matrix between all the patients using the function, JSD_matrix.
}
